fix: use SQLAlchemy Integer type instead of Python Decimal in cast
Fixed AttributeError when exporting Excel report.
Changed func.cast(ParticipantPayment.is_paid, Decimal) to use Integer type
which is the correct SQLAlchemy type for casting boolean to numeric.

Error was: AttributeError: 'decimal.Decimal' object has no attribute '_isnull'

+++ app/routers/dashboard.py
@@ -408,9 +408,11 @@ async def export_project_excel(
         func.count(Expense.id).label("count"),
     ).filter(Expense.project_id == project.id).first()
 
+    from sqlalchemy import Integer
+
     payment_totals = db.query(
         func.count(ParticipantPayment.id).label("total_payments"),
-        func.sum(func.cast(ParticipantPayment.is_paid, Decimal)).label("paid_count"),
+        func.sum(func.cast(ParticipantPayment.is_paid, Integer)).label("paid_count"),
     ).join(Expense).filter(Expense.project_id == project.id).first()
 
     total_usd = float(expense_totals.total_usd or 0)
